Added variant.h/cpp; changed variable names in object.h; added some functions to wxStringList; some bugs fixes
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@730 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/common/prntbase.cpp
@@ -613,7 +613,7 @@ bool wxPrintPreviewBase::PaintPage(wxWindow *canvas, wxDC& dc)
   double actualWidth = (zoomScale*m_pageWidth*m_previewScale);
 //  float actualHeight = (float)(zoomScale*m_pageHeight*m_previewScale);
 
-  int x = ((canvasWidth - actualWidth)/2.0);
+  int x = (int) ((canvasWidth - actualWidth)/2.0);
   if (x < m_leftMargin)
     x = m_leftMargin;
   int y = m_topMargin;
